Quantify protein concentration from a MicroBCA assay

Quantify protein concentration from a MicroBCA assay

Usage

qp(
  x,
  replicate_orientation = c("h", "v"),
  sample_names = NULL,
  remove_empty = TRUE,
  ignore_outliers = c("all", "samples", "standards", "none"),
  standard_scale = c(0, 2^((2:7) - 5)),
  n_replicates = 3,
  wavelength = 562
)

Arguments

x

A spectramax, gp, or data.frame object, or path to SPECTRAmax .xls(x)/.txt file.

replicate_orientation

Either 'h' or 'v' - see Details.

sample_names

Optional character vector of sample names.

remove_empty

Should wells that have less absorbance than the lowest standard be dropped?

ignore_outliers

Character. From which group - samples or standards - should outliers be detected and removed?

standard_scale

Numeric. Known concentrations of standards, in the order they appear.

n_replicates

Numeric. The number of techinical replicates.

wavelength

Numeric. The wavelength absorbance was captured.

Value

a tibble

Details

If x is a spectramax, the standards must start in the upper left corner in the order dictated by standard_scale. Whether this is from from left to right or top to bottom can be specified in replicate_orientation. Note that replicate_orientation specified the direction that REPLICATES lie, NOT the direction the samples flow (which will be perpendicular to the replicates).

Note: replicate_orientation, n_replicates, and wavelength will be silently ignored if x is not a spectramax or path to a spectramax

Examples

data <- system.file("extdata", "absorbances.txt", package = "qp")
qp(data, replicate_orientation = "h")
#> $fit
#> 
#> Call:
#> stats::lm(formula = .log2_conc ~ .log2_abs, data = fit_data)
#> 
#> Coefficients:
#> (Intercept)    .log2_abs  
#>       3.358        1.141  
#> 
#> 
#> $qp
#> # A tibble: 36 × 13
#>     .row  .col   .abs sample_type index .conc .is_outlier  .mean .log2_abs
#>    <int> <dbl>  <dbl> <fct>       <dbl> <dbl> <lgl>        <dbl>     <dbl>
#>  1     1     1 0.0686 standard        1 0     FALSE       0.0700     -3.87
#>  2     1     2 0.0717 standard        1 0     FALSE       0.0700     -3.80
#>  3     1     3 0.0698 standard        1 0     FALSE       0.0700     -3.84
#>  4     2     1 0.0825 standard        2 0.125 TRUE        0.0834     -3.60
#>  5     2     2 0.0832 standard        2 0.125 FALSE       0.0834     -3.59
#>  6     2     3 0.0835 standard        2 0.125 FALSE       0.0834     -3.58
#>  7     3     1 0.102  standard        3 0.25  FALSE       0.102      -3.29
#>  8     3     2 0.100  standard        3 0.25  TRUE        0.102      -3.32
#>  9     3     3 0.102  standard        3 0.25  FALSE       0.102      -3.30
#> 10     4     1 0.132  standard        4 0.5   FALSE       0.132      -2.92
#> # ℹ 26 more rows
#> # ℹ 4 more variables: .pred <dbl>, .pred_conc <dbl>, .pred_conc_mean <dbl>,
#> #   .sample_name <chr>
#>